requirements.txt
google-adk
google-adk[a2a]
a2a-sdk